+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 0.14.15 (tbd)
+
+- Fixed issue with `get-dependency-map` of the `piral-debug-utils`
+- Added support for v0 and v1 pilets in `get-dependency-map`
+- Added support for object flags (e.g., `--vars`) in the initializers
+
 ## 0.14.14 (March 3, 2022)
 
 - Fixed issue placing *tsconfig.json* in empty template